# Switch the AST over to Cow<str>