--- a/fluent-syntax/src/ast.rs
+++ b/fluent-syntax/src/ast.rs
@@ -1,3 +1,5 @@
+use std::borrow::Cow;
+
 #[derive(Debug, PartialEq)]
 pub struct Resource<'ast> {
     pub body: Vec<ResourceEntry<'ast>>,
@@ -6,7 +8,7 @@ pub struct Resource<'ast> {
 #[derive(Debug, PartialEq)]
 pub enum ResourceEntry<'ast> {
     Entry(Entry<'ast>),
-    Junk(&'ast str),
+    Junk(Cow<'ast, str>),
 }
 
 #[derive(Debug, PartialEq)]
@@ -39,7 +41,7 @@ pub struct Pattern<'ast> {
 
 #[derive(Debug, PartialEq)]
 pub enum PatternElement<'ast> {
-    TextElement(&'ast str),
+    TextElement(Cow<'ast, str>),
     Placeable(Expression<'ast>),
 }
 
@@ -51,7 +53,7 @@ pub struct Attribute<'ast> {
 
 #[derive(Debug, PartialEq)]
 pub struct Identifier<'ast> {
-    pub name: &'ast str,
+    pub name: Cow<'ast, str>,
 }
 
 #[derive(Debug, PartialEq)]
@@ -63,24 +65,24 @@ pub struct Variant<'ast> {
 
 #[derive(Debug, PartialEq)]
 pub enum VariantKey<'ast> {
-    Identifier { name: &'ast str },
-    NumberLiteral { value: &'ast str },
+    Identifier { name: Cow<'ast, str> },
+    NumberLiteral { value: Cow<'ast, str> },
 }
 
 #[derive(Debug, PartialEq)]
 pub enum Comment<'ast> {
-    Comment { content: Vec<&'ast str> },
-    GroupComment { content: Vec<&'ast str> },
-    ResourceComment { content: Vec<&'ast str> },
+    Comment { content: Vec<Cow<'ast, str>> },
+    GroupComment { content: Vec<Cow<'ast, str>> },
+    ResourceComment { content: Vec<Cow<'ast, str>> },
 }
 
 #[derive(Debug, PartialEq)]
 pub enum InlineExpression<'ast> {
     StringLiteral {
-        value: &'ast str,
+        value: Cow<'ast, str>,
     },
     NumberLiteral {
-        value: &'ast str,
+        value: Cow<'ast, str>,
     },
     FunctionReference {
         id: Identifier<'ast>,