Fixes #104