@Michael-F-Bryan 2020-06-23
Benchmark results after replacing `&'ast str` with `Cow<'ast, str>`:

```console
$ cargo bench --bench parser -- --baseline base
   Compiling fluent-syntax v0.9.3 (C:\Projects\fluent-rs\fluent-syntax)
    Finished bench [optimized] target(s) in 9.46s
     Running C:\Projects\fluent-rs\target\release\deps\parser-99e6937ba38adfc9.exe
Gnuplot not found, using plotters backend
parse/"simple"          time:   [43.395 us 45.529 us 48.144 us]
                        change: [+25.007% +29.004% +33.622%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 8 outliers among 100 measurements (8.00%)
  5 (5.00%) high mild
  3 (3.00%) high severe
parse/"preferences"     time:   [387.97 us 390.36 us 393.46 us]
                        change: [+3.4989% +6.0201% +8.2245%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe
parse/"menubar"         time:   [104.65 us 108.02 us 112.16 us]
                        change: [+6.1007% +8.5359% +11.820%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 10 outliers among 100 measurements (10.00%)
  4 (4.00%) high mild
  6 (6.00%) high severe

unicode                 time:   [7.7112 us 8.0368 us 8.4848 us]
                        change: [+1.5234% +4.8896% +8.6534%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 12 outliers among 100 measurements (12.00%)
  3 (3.00%) high mild
  9 (9.00%) high severe

parse_ctx/"browser"     time:   [399.25 us 405.31 us 412.26 us]
                        change: [+5.0663% +6.1496% +7.3562%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
Benchmarking parse_ctx/"preferences": Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.1s or reduce sample count to 70.
parse_ctx/"preferences" time:   [989.68 us 1.0161 ms 1.0505 ms]
                        change: [+8.0732% +11.960% +16.002%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
```

Other than the "simple" benchmark, we're looking at a 6-11% performance regression.