@zbraniecki 2020-06-23
I'll measure against Fx, but `6-11%` would likely translate to ~1.5-2ms on startup on a very fast machine - https://bugzilla.mozilla.org/show_bug.cgi?id=1560038#c57

On a slower user machine this may be 5-10ms. I'd prefer to maintain a ready-only AST over paying that price for convenience.