@Michael-F-Bryan 2020-06-23
Benchmark comparing 59fb6ad against master (c9e4565) after updating `fluent-bundle` to use `Cow<str>`.

```console
$ cargo bench --bench resolver -- --baseline base
   Compiling fluent-syntax v0.9.3 (C:\Projects\fluent-rs\fluent-syntax)
   Compiling fluent-bundle v0.11.0 (C:\Projects\fluent-rs\fluent-bundle)
    Finished bench [optimized] target(s) in 11.11s
     Running C:\Projects\fluent-rs\target\release\deps\resolver-d65a2ee18bbb4c2c.exe
Gnuplot not found, using plotters backend
resolve/"simple"        time:   [9.5628 us 9.6640 us 9.7876 us]
                        change: [-14.508% -10.849% -7.4316%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe
resolve/"preferences"   time:   [140.55 us 142.14 us 144.46 us]
                        change: [+2.1204% +5.5609% +8.9761%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
resolve/"menubar"       time:   [31.260 us 31.382 us 31.524 us]
                        change: [-2.0665% -0.6816% +0.5605%] (p = 0.33 > 0.05)
                        No change in performance detected.
resolve/"unescape"      time:   [1.7500 us 1.7625 us 1.7773 us]
                        change: [-2.0425% -0.9417% -0.0567%] (p = 0.06 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
```

It looks like the "preferences" benchmark took a 5% hit, so switching to `Cow<str>` also has a negative effect when formatting a message.

> I'd prefer to maintain a ready-only AST over paying that price for convenience.

@zbraniecki, I agree. I was thinking we could use [Yandros's trick](https://users.rust-lang.org/t/fluent-l10n-system-parser-ast-serializer-decision/27507/2?u=michael-f-bryan) of parameterising the AST nodes over some type `S: AsRef<str>` and giving it a default value of `S = &'ast str`. That way the base case will be identical to the current AST, but if you need to do mutations you can transform it to using `Cow<str>` when folding over the tree... Thoughts?