@Michael-F-Bryan 2020-07-05
Here are the benchmark results for #187 (using an `ArcStr` instead of `&str` or `Cow<str>`) and you can see performance regressed by a non-trivial amount. 

I thought it might have been due to the way the benchmarks are constructed (it stores the text as `String`s, so every run of every benchmark ended up copying the contents to a new `ArcStr`), but switching everything to `ArcStr` in Michael-F-Bryan/fluent-rs@15d0d8527a404c419aa5c096ae5428ee424eff41 to skip the initial copy didn't help too much...

```console
# Baseline (master)

$ git checkout master && git rev-parse HEAD
c9e456515f4c084eb9650f3910722a31707a6b32
$ cargo bench --bench parser
parse/"simple"          time:   [12.823 us 12.903 us 13.024 us]
Found 14 outliers among 100 measurements (14.00%)
  9 (9.00%) high mild
  5 (5.00%) high severe
parse/"preferences"     time:   [236.38 us 238.11 us 240.29 us]
Found 15 outliers among 100 measurements (15.00%)
  5 (5.00%) low mild
  7 (7.00%) high mild
  3 (3.00%) high severe
parse/"menubar"         time:   [58.148 us 58.343 us 58.602 us]
Found 22 outliers among 100 measurements (22.00%)
  4 (4.00%) low severe
  1 (1.00%) high mild
  17 (17.00%) high severe

unicode                 time:   [4.6397 us 4.6947 us 4.7749 us]
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) high mild
  14 (14.00%) high severe

parse_ctx/"browser"     time:   [245.25 us 247.41 us 250.85 us]
Found 10 outliers among 100 measurements (10.00%)
  8 (8.00%) high mild
  2 (2.00%) high severe
parse_ctx/"preferences" time:   [604.51 us 606.86 us 609.26 us]

# With ArcStr

$ git checkout rc-str-ast && git rev-parse HEAD
e807455276524572f69d9f275dbff005d6928d31
$ cargo bench --bench parser
parse/"simple"          time:   [23.790 us 24.533 us 25.135 us]
                        change: [+65.766% +70.315% +75.086%] (p = 0.00 < 0.05)
                        Performance has regressed.
parse/"preferences"     time:   [257.25 us 258.34 us 259.64 us]
                        change: [+7.8525% +8.9906% +10.152%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 29 outliers among 100 measurements (29.00%)
  9 (9.00%) low severe
  4 (4.00%) high mild
  16 (16.00%) high severe
parse/"menubar"         time:   [72.983 us 73.360 us 73.747 us]
                        change: [+24.247% +25.545% +26.511%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 37 outliers among 100 measurements (37.00%)
  3 (3.00%) low severe
  16 (16.00%) low mild
  10 (10.00%) high mild
  8 (8.00%) high severe

unicode                 time:   [4.6590 us 4.6968 us 4.7400 us]
                        change: [-1.2845% +0.2354% +1.3430%] (p = 0.78 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

parse_ctx/"browser"     time:   [277.00 us 280.18 us 283.69 us]
                        change: [+10.566% +11.847% +13.102%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe
parse_ctx/"preferences" time:   [655.87 us 662.29 us 670.78 us]
                        change: [+7.6288% +8.5293% +9.5709%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high severe

# Tweaked the benchmarks to not copy the source text every time

$ git rev-parse HEAD
15d0d8527a404c419aa5c096ae5428ee424eff41
$ cargo bench --bench parser
parse/"simple"          time:   [19.246 us 19.453 us 19.764 us]
                        change: [-14.504% -11.978% -9.3610%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
parse/"preferences"     time:   [252.80 us 257.30 us 262.51 us]
                        change: [-2.6368% -1.7249% -0.7337%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 38 outliers among 100 measurements (38.00%)
  21 (21.00%) low severe
  1 (1.00%) low mild
  2 (2.00%) high mild
  14 (14.00%) high severe
parse/"menubar"         time:   [71.020 us 71.441 us 71.934 us]
                        change: [-3.5266% -2.9758% -2.3899%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe

unicode                 time:   [4.7729 us 4.8176 us 4.8938 us]
                        change: [+1.9107% +3.4854% +5.6462%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 16 outliers among 100 measurements (16.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
  12 (12.00%) high severe

parse_ctx/"browser"     time:   [270.87 us 272.46 us 274.06 us]
                        change: [-2.6351% -1.7316% -0.8859%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
parse_ctx/"preferences" time:   [641.53 us 646.67 us 652.89 us]
                        change: [-3.7101% -2.6661% -1.8317%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high severe
```

If it were just a case of nice code I'd prefer the `ArcStr` AST over any of the other solutions in this PR, simply because not having to worry about lifetime annotations or use `rental` to manage the string's memory makes the code more ergonomic... But considering Fluent is used by Firefox to localise the UI and parsing directly impacts Firefox's startup time, I'm inclined to go with the generic AST + `Visitor<S>` approach.